iox-#1230 Move barrier to testing
